move tests to correct homes since their classes moved
